Add more, not reproducible benchmarks
The benchmarks are currently hardcoded to use a specific system
on my computer. To run the tests yourself, change QUERY_DERIV__OLD in
benches/common.rs

In the future, I would like to pull a specific copy of the
nixos sqlite db from e.g. a github repo to make the benchmarks
reproducible